# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  app_internal_interfaces_http_dto_request.ContactMessageStatusUpdateRequest:
    properties:
      status:
        enum:
        - unread
        - read
        - archived
        type: string
    required:
    - status
    type: object
  app_internal_interfaces_http_dto_request.ContactMessageSubmitRequest:
    properties:
      email:
        maxLength: 255
        type: string
      message:
        type: string
      name:
        maxLength: 255
        type: string
      phone:
        maxLength: 50
        type: string
      source:
        enum:
        - homepage
        - about
        - contact_page
        type: string
      subject:
        maxLength: 500
        type: string
    required:
    - email
    - message
    - name
    type: object
...
  app_internal_interfaces_http_dto_response.ContactMessageListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/app_internal_interfaces_http_dto_response.ContactMessageResponse'
        type: array
      meta:
        $ref: '#/definitions/app_pkg_pagination.Meta'
    type: object
  app_internal_interfaces_http_dto_response.ContactMessageResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      message:
        type: string
      name:
        type: string
      phone:
        type: string
      source:
        type: string
      status:
        type: string
      subject:
        type: string
    type: object
...
  app_internal_interfaces_http_dto_response.ErrorBody:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        example: request validation failed
        type: string
    type: object
  app_internal_interfaces_http_dto_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorBody'
      path:
        example: /api/v1/auth/login
        type: string
      statusCode:
        example: 400
        type: integer
    type: object
...
  app_pkg_pagination.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@shop.com
    name: API Support
  description: REST API for the eCommerce admin panel and public storefront. Supports
    JWT authentication with role-based access (admin, customer). Store endpoints under
    /api/v1/store are public; account routes require a customer token.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: eCommerce API
  version: "1.0"
paths:
...
  /api/v1/admin/contact-messages:
    get:
      description: Get a paginated list of contact messages.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Status filter (unread/read/archived)
        in: query
        name: status
        type: string
      - description: Source filter (homepage/about/contact_page)
        in: query
        name: source
        type: string
      - description: Search query by name, email, or subject
        in: query
        name: q
        type: string
      - description: Filter messages created on or after date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Filter messages created on or before date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ContactMessageListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List contact messages (Admin)
      tags:
      - admin-contact
  /api/v1/admin/contact-messages/{id}:
    delete:
      description: Delete a contact message.
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete contact message (Admin)
      tags:
      - admin-contact
    get:
      description: Get details of a single contact message.
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ContactMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get contact message (Admin)
      tags:
      - admin-contact
  /api/v1/admin/contact-messages/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update status of a contact message.
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app_internal_interfaces_http_dto_request.ContactMessageStatusUpdateRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update contact status (Admin)
      tags:
      - admin-contact
...
  /api/v1/store/contact:
    post:
      consumes:
      - application/json
      description: Submit a contact inquiry form.
      parameters:
      - description: Contact form data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app_internal_interfaces_http_dto_request.ContactMessageSubmitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ContactMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      summary: Submit contact message
      tags:
      - store-contact
...
securityDefinitions:
  BearerAuth:
    description: 'Enter "Bearer" followed by a space and the JWT access token. Example:
      Bearer eyJhbGciOiJIUzI1NiIs...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"